feat: Change Enum Path and Add Name
- Change Enums Path
- Add Enum Name

#10
