--- src/main/java/com/account/yomankum/domain/Gender.java
@@ -1,12 +1,0 @@
-package com.account.yomankum.domain;
-
-public enum Gender {
-
-    MALE("남자"), FEMALE("여자");
-
-    private String name;
-
-    Gender(String name) {
-        this.name = name;
-    }
-}
