--- src/main/java/com/account/yomankum/domain/Mail.java
@@ -1,5 +1,0 @@
-package com.account.yomankum.domain;
-
-public enum Mail {
-    JOIN
-}
