+++ src/main/java/com/account/yomankum/domain/enums/Job.java
@@ -1,0 +1,12 @@
+package com.account.yomankum.domain.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum Job {
+    EX("기타");
+
+    private String job;
+}
