+++ src/main/java/com/account/yomankum/domain/enums/Mail.java
@@ -1,0 +1,5 @@
+package com.account.yomankum.domain.enums;
+
+public enum Mail {
+    JOIN
+}
