+++ src/main/java/com/account/yomankum/domain/enums/Name.java
@@ -1,4 +1,4 @@
-package com.account.yomankum.domain;
+package com.account.yomankum.domain.enums;
 
 public enum Name {
     ROLE_USER, ROLE_ADMIN
