+++ src/main/java/com/account/yomankum/login/service/MailService.java
@@ -1,6 +1,6 @@
 package com.account.yomankum.login.service;
 
-import com.account.yomankum.domain.Mail;
+import com.account.yomankum.domain.enums.Mail;
 import jakarta.mail.Message;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
