+++ src/main/java/com/account/yomankum/login/service/SnsUserServiceImpl.java
@@ -1,7 +1,7 @@
 package com.account.yomankum.login.service;
 
 import com.account.yomankum.security.oauth.Sns;
-import com.account.yomankum.domain.Name;
+import com.account.yomankum.domain.enums.Name;
 import com.account.yomankum.domain.Role;
 import com.account.yomankum.domain.SnsUser;
 import com.account.yomankum.repository.SnsUserRepository;
