+++ src/main/java/com/account/yomankum/login/service/UserServiceImpl.java
@@ -1,7 +1,7 @@
 package com.account.yomankum.login.service;
 
 import com.account.yomankum.security.jwt.TokenService;
-import com.account.yomankum.domain.Name;
+import com.account.yomankum.domain.enums.Name;
 import com.account.yomankum.domain.Role;
 import com.account.yomankum.domain.User;
 import com.account.yomankum.login.domain.LoginDto;
