+++ src/main/java/com/account/yomankum/security/jwt/TokenService.java
@@ -1,6 +1,6 @@
 package com.account.yomankum.security.jwt;
 
-import com.account.yomankum.domain.Name;
+import com.account.yomankum.domain.enums.Name;
 
 public interface TokenService {
